refactor: 시작페이지 고정 버튼 코드분리
