+++ src/pages/Start/StartAnimation/GiftAnimation/StartGiftAnimation.style.ts
