+++ src/pages/Start/StartAnimation/ScrollAnimation/StartScrollAnimation.style.ts
