+++ src/pages/Start/StartFixedButton/StartFixedButton.style.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import { IcKakoLarge } from '../../../assets/svg';
+
+export const KakaoLogin = styled(IcKakoLarge)`
+  position: fixed;
+  bottom: 0;
+  z-index: 3;
+  width: 33.5rem;
+  margin: 2rem;
+`;
